+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Toolium Changelog
 =================
 
+v2.4.1
+------
+
+*Release date: In development*
+
 v2.4.0
 ------